mejorando la funcion convertir ruta relativa a absoluta
